optimisations and bugs fix

+++ src/main/java/data/PVars.java
@@ -6,7 +6,7 @@ import arc.struct.Seq;
 public class PVars {
 	public static Seq<String> adminCommands = new Seq<>();
     public static mindustry.maps.Map selectedMap;
-    public static byte waveVoted = 1;
+    public static short waveVoted = 1;
     public static boolean tchat = true,
     	autoPause = false, 
     	niceWelcome = true,
